+++ CHANGELOG.md
@@ -10,6 +10,7 @@ List all changes after the last release here (newer on top). Each change on a se
 
 ### Changed
 
+- Xtheme: allow async plugin to have orderable only flag set
 - Xtheme: use the current language as part of the plugins cache key
 
 ## [2.8.3] - 2021-05-17