+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Fixed
+
+- Use more bites in order to detect csv dialect in importer
+
 ## [2.8.3] - 2021-05-17
 
 ### Fixed