+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+## [2.9.1] - 2021-05-24
+
 ### Fixed
 
 - Admin: add ordering to attribute filter, because of frontend related error